option to use either hardware UART or software bitbang implementation
